Initial Commit of new Fit.
A mayor refactoring has been made here, which seperates Objective
functions from the Minimizer. This should make it easier to extend
symfit with new Minimizers in particular, but also with other objects
and functionality.

On top of this, Fit is now a lot smater and will attempt to choose the
right strategy for your problem.

MINPACK is no longer the default, but it is still available for
backwardscompatibility.