Introduce ScipyConstrainedMinimize.
This brings constraints on the same footing as jacobians by bringing the conversion to scipy constraints to a seperate subclass. Inherites from ScipyGradientMinimize because we need this to wrap the jacobian of the constraints in order to feed them to SLSQP, but this might be factored out because scipy's COBYLA does not accept a jacobian for it's constraints.