Made Model.chi_squared functional for IndexedParameters.
Additionally, removed Model.chi, Model.chi_jacobian, and its numerical counterparts. This was a remnant from a dark past when we still had to bow to the will of MINPACK.

Removed without deprication, as I seriously doubt anybody even knew about this. Symfit itself didn't even use it internally or test for it.