Add __version__ (#167)
Fixed #35 by getting the version number from pkg_resources.

This will also solve the problems with the documentation version, since the stable doc build is the default, and those are only triggered when a new version is tagged. This version can then be correctly read by pkg_recources when building the docs.