Updated the tests to the new covariance matrix computation.
Tests using a GradientModel will give the same result as the previous test, and so this is used for testing e.g. ODEModels or comparing to Mathematica.
For some tests where we prefer to keep a Model we need to make the tests less strict, since the exact Hessian will give a slightly different result than the traditional approximation.