Removed the default setting of variables to 'None' if no data is provided.
Now that we have a proper connectivity_mapping, it is easy to deduce which variables need data and which don't. This fixes a slew of Heisenbugs which I have come across when fitting certain problems, where you might forgot to provide the correct data to a variable but symfit does not raise an error.

This fixes #214. It does make slight API changes under certain conditions and therefore requires some tests to be modified, but the changes are well worth it and self explanatory due to the precise nature of the errors raised.