documentation improvements (#306)
* doc/html: fix falsely set static path

sphinx will automatically copy the referenced images to _images, so setting the directory will result in duplicate files.

Signed-off-by: Stephan Lachnit <stephanlachnit@protonmail.com>

* doc/html: don't copy rst sources

By default the rst sources are added to the html build, however that doesn't add much value. Also, the paths to the referenced images from the rst files are incorrect.

Signed-off-by: Stephan Lachnit <stephanlachnit@protonmail.com>

* doc/html: don't set a theme

Not setting a theme will result in sphinx using its recommended theme. Interestingly this is different from the 'default' theme, which in fact is the classic theme.

Signed-off-by: Stephan Lachnit <stephanlachnit@protonmail.com>

* docs/conf.py: document all variables

Signed-off-by: Stephan Lachnit <stephanlachnit@protonmail.com>

* doc/html: document mathjax_path

By default, the html build will include a link to the mathjax script, however one might want to set the path to an offline installation of mathjax.
The commented path works for the standard installation in Debian- and Fedora-based distributions.

Signed-off-by: Stephan Lachnit <stephanlachnit@protonmail.com>

* docs/Makefile: set doctreesdir outside of builddir

With this, docs/_build basically only contains files that should actually be installed (in the html target), making it easier to deploy. Also, a rebuild can be triggered without deleting the doctrees chache.

Signed-off-by: Stephan Lachnit <stephanlachnit@protonmail.com>

* gitignore: housekeeping

docs_build isn't actually used, and docs/_doctrees is added for 9a5b459b1e13eb06afc7eb7a612c3f5df6374407.

Signed-off-by: Stephan Lachnit <stephanlachnit@protonmail.com>

* doc/html: document nbsphinx_requirejs_path

By default, the html build will include a link to a requirejs script, however one might want to set the path to an offline installation of requirejs.
The commented path works for the standard installation in Debian- and Fedora-based distributions.

Signed-off-by: Stephan Lachnit <stephanlachnit@protonmail.com>